// frontend/src/generated/backend/models/FailureReport.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { GithubIssue } from './GithubIssue';
export type FailureReport = {
    id?: number;
    errorName: string;
    errorDetail: string;
    issueNumber?: number;
    createDatetime: string;
    githubIssue?: GithubIssue;
};